Split up load and save tests into their own files
